# docs/contracts/api-spec.yaml
openapi: "3.0.3"
info:
  title: 租车应用 API
  version: "2.0"
  description: 租车小程序 + PC 管理端 RESTful API
servers:
  - url: /api/v1
    description: v1 API
security:
  - BearerAuth: []
paths:
...
  /auth/wx-login:
    post:
      summary: 微信小程序登录
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WxLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '4001':
          description: 微信登录失败
        '4002':
          description: 手机号获取失败
  /auth/admin-login:
    post:
      summary: PC 管理端登录
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: 登录成功
...
  /vehicles/{id}:
    get:
      summary: 车辆详情
      tags: [Vehicle]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 车辆详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/VehicleDetail'
        '4004':
          description: 车辆不存在

  # ==================== 订单模块（用户端） ====================
  /orders:
    post:
      summary: 创建订单
      tags: [Order]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: 订单创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreateOrderResult'
        '5200':
          description: 时间段冲突
        '4004':
          description: 车辆不存在或已下架
    get:
      summary: 我的订单列表
      tags: [Order]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: 订单列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
  /orders/{id}:
    get:
      summary: 订单详情
      tags: [Order]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderDetail'
        '4004':
          description: 订单不存在
  /orders/{id}/cancel:
    post:
      summary: 取消订单
      tags: [Order]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 取消成功
        '5300':
          description: 订单状态不允许取消
...
  /admin/orders:
    get:
      summary: 订单列表（管理端）
      tags: [Admin Order]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: vehicle_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: 订单列表
  /admin/orders/{id}/confirm:
    post:
      summary: 确认订单
      tags: [Admin Order]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 确认成功
        '5300':
          description: 订单状态不允许确认
  /admin/orders/{id}/reject:
    post:
      summary: 拒绝订单
      tags: [Admin Order]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectOrderRequest'
      responses:
        '200':
          description: 拒绝成功
        '5300':
          description: 订单状态不允许拒绝
  /admin/orders/{id}/start:
    post:
      summary: 标记订单进行中
      tags: [Admin Order]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 标记成功
        '5300':
          description: 订单状态不允许开始
  /admin/orders/{id}/complete:
    post:
      summary: 标记订单完成
      tags: [Admin Order]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 标记成功
        '5300':
          description: 订单状态不允许完成

  # ==================== 管理端 - 车辆 ====================
  /admin/vehicles:
    get:
      summary: 车辆列表（管理端，含下架）
      tags: [Admin Vehicle]
      responses:
        '200':
          description: 车辆列表
    post:
      summary: 创建车辆
      tags: [Admin Vehicle]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
      responses:
        '200':
          description: 创建成功
  /admin/vehicles/{id}:
    put:
      summary: 更新车辆
      tags: [Admin Vehicle]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
      responses:
        '200':
          description: 更新成功
        '4004':
          description: 车辆不存在
    delete:
      summary: 删除车辆（软删除）
      tags: [Admin Vehicle]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
  /admin/vehicles/{id}/toggle-status:
    post:
      summary: 上下架切换
      tags: [Admin Vehicle]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 状态切换成功
        '4004':
          description: 车辆不存在

  # ==================== 管理端 - 价格设置 ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
      description: 页码
...
  schemas:
...
    ApiResponse:
      type: object
      required: [code, data, message]
      properties:
        code:
          type: integer
          description: 错误码，0 表示成功
        data:
          type: object
          nullable: true
          description: 响应数据
        message:
          type: string
          description: 响应消息

    # ===== 认证相关 =====
    WxLoginRequest:
      type: object
      required: [login_code, phone_code]
      properties:
        login_code:
          type: string
          description: wx.login() 获取的 code
        phone_code:
          type: string
          description: getUserPhoneNumber 获取的 code
    AdminLoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          description: 管理端登录密码
...
    VehicleDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        brand:
          type: string
        seats:
          type: integer
        transmission:
          type: string
          enum: [auto, manual]
        description:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
        weekday_price:
          type: string
          format: decimal
        weekend_price:
          type: string
          format: decimal
        holiday_price:
          type: string
          format: decimal
          nullable: true
        tags:
          type: array
          items:
            type: string
...
    CreateOrderRequest:
      type: object
      required: [vehicle_id, start_date, end_date]
      properties:
        vehicle_id:
          type: integer
          format: int64
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    PriceBreakdownItem:
      type: object
      properties:
        date:
          type: string
          format: date
        type:
          type: string
          enum: [weekday, weekend, holiday]
        price:
          type: string
          format: decimal
    CreateOrderResult:
      type: object
      properties:
        order_id:
          type: integer
          format: int64
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        days:
          type: integer
        total_price:
          type: string
          format: decimal
    OrderListItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        vehicle_name:
          type: string
        vehicle_image:
          type: string
          description: 取自车辆 images[0]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        total_price:
          type: string
          format: decimal
        status:
          $ref: '#/components/schemas/OrderStatus'
        created_at:
          type: string
          format: date-time
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        vehicle:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
            images:
              type: array
              items:
                type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        days:
          type: integer
        total_price:
          type: string
          format: decimal
        status:
          $ref: '#/components/schemas/OrderStatus'
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        price_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/PriceBreakdownItem'
        created_at:
          type: string
          format: date-time
    OrderListResponse:
      type: object
      properties:
        code:
          type: integer
        data:
          type: object
          properties:
            total:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderListItem'
    RejectOrderRequest:
      type: object
      properties:
        reason:
          type: string
          description: 拒绝原因（可选）

    # ===== 订阅消息 =====
...
    CreateVehicleRequest:
      type: object
      required: [name, brand, seats, transmission, weekday_price, weekend_price]
      properties:
        name:
          type: string
        brand:
          type: string
        seats:
          type: integer
        transmission:
          type: string
          enum: [auto, manual]
        description:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
        weekday_price:
          type: string
          format: decimal
        weekend_price:
          type: string
          format: decimal
        holiday_price:
          type: string
          format: decimal
          nullable: true
          description: 不填时默认使用 weekend_price

    # ===== 价格设置 =====
...
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - in_progress
        - completed
        - cancelled
        - rejected
      description: |
        pending: 待确认
        confirmed: 已确认
        in_progress: 进行中
        completed: 已完成
        cancelled: 已取消
        rejected: 已拒绝
    PaymentStatus:
      type: string
      enum:
        - unpaid
        - paid
        - refunded
      description: |
        unpaid: 未支付
        paid: 已支付
        refunded: 已退款